+++ src/extension.ts
@@ -1,9 +1,10 @@
 import * as vscode from 'vscode';
 import { Utils as UriUtils } from 'vscode-uri';
+import * as os from 'os';
 import { Template } from './schemas';
 import { createTemplateEditAsync } from './templateEdit';
-import { getManifestPath, getRegisteredTemplates, loadManifestAsync, registerTemplate, unregisterTemplate } from './templates';
-import { showTemplateWizardAsync } from './templatesUI';
+import { getManifestPath as getManifestFileUri, getRegisteredTemplates, getWorkspaceManifestPath as getWorkspaceManifestFileUri, loadManifestAsync, registerTemplate, unregisterTemplate } from './templates';
+import { showTemplateWizardAsync, TemplateRootUriTuple } from './templatesUI';
 
 /**
  * Extension startup. 
@@ -13,7 +14,16 @@ export function activate(context: vscode.ExtensionContext) {
   context.subscriptions.push(
     vscode.commands.registerCommand(
       'templates.newItem',
-      (folder?: vscode.Uri | string): Promise<void> => createNewItemsAsync(folder)
+      async (folderPathOrUri?: vscode.Uri | string): Promise<void> => {
+        // Workspace folder is not be available in all contexts,
+        // See: https://github.com/Microsoft/vscode/issues/3553
+        const folder = folderPathOrUri ?? vscode.workspace.workspaceFolders?.at(0)?.uri;
+        if (!folder) {
+          vscode.window.showErrorMessage("No target folder to add items to");
+          return;
+        }
+        await createNewItemsAsync(folder);
+      }
     ),
     vscode.commands.registerCommand(
       'templates.registerTemplate',
@@ -30,45 +40,25 @@ export function activate(context: vscode.ExtensionContext) {
  * Add a new item using a template.
  * 
  * @param folder Folder to add new item to. Must be an absolute URI or path to
- * a folder that belongs to an open workspace. If missing, root of currently
- * opened workspace will be used.
+ * a folder that belongs to an open workspace.
  */
-async function createNewItemsAsync(folder?: vscode.Uri | string): Promise<void> {
-  // Workspace folder might not be available in all contexts.
-  // See: https://github.com/Microsoft/vscode/issues/3553
-  const folderUri = folder
-    ? folder instanceof vscode.Uri ? folder : vscode.Uri.parse(folder, true)
-    : vscode.workspace.workspaceFolders?.at(0)?.uri;
+async function createNewItemsAsync(folder: vscode.Uri | string): Promise<void> {
+  const targetFolderUri = folder instanceof vscode.Uri ? folder : vscode.Uri.parse(folder, true);
 
-  if (!folderUri) {
-    throw new Error("BUG: Missing required target folder.");
-  }
-
-  const workspaceFolder = vscode.workspace.getWorkspaceFolder(folderUri);
-  if (!workspaceFolder) {
-    vscode.window.showErrorMessage(`Target folder does not belong to a currently open workspace. Folder:${folderUri}`);
+  const targetWorkspaceFolder = vscode.workspace.getWorkspaceFolder(targetFolderUri);
+  if (!targetWorkspaceFolder) {
+    vscode.window.showErrorMessage(
+      `Target folder does not belong to a currently open workspace. Folder:${targetFolderUri}`);
     return;
   }
 
-  const templates = new Map<string, Template>(getRegisteredTemplates());
-
-  // Load templates from workspace every time and do not register them to avoid
-  // the complexities of monitoring changes on workspace settings or the FS.
-  const manifestUri = getManifestPath(workspaceFolder);
-  const manifestExists = await vscode.workspace.fs.stat(manifestUri).then(() => true, () => false);
-  if (manifestExists) {
-    const manifest = await loadManifestAsync(manifestUri);
-    manifest.templates.forEach((value: Template, index: number) => {
-      templates.set(`${value.name}@${manifestUri}[${index}]`, value);
-    });
-  }
-
+  // Load templates every time to avoid the risks and complexities of monitoring 
+  // the FS on a feature that would be uncommonly used. 
+  const templates = await loadTemplates(targetWorkspaceFolder);
   if (templates.size === 0) {
-    const friendlyLocation = vscode.workspace.asRelativePath(manifestUri, true)
-    const msg = `There are no templates available in workspace '${workspaceFolder.name}'. They can be defined in ${friendlyLocation}`;
-
-    // TODO: Offer to create a template file.
-    vscode.window.showWarningMessage(msg);
+    // TODO: Offer to create a template file?
+    vscode.window.showWarningMessage(
+      "There are no templates available. They can be define in your workspace or user directory");
     return;
   }
 
@@ -77,13 +67,14 @@ async function createNewItemsAsync(folder?: vscode.Uri | string): Promise<void> 
     return; // cancelled by user
   }
 
-  const templateRootUri = vscode.Uri.joinPath(UriUtils.dirname(manifestUri), selection.template.location);
+  const rootUri = selection.rootUri ?? targetWorkspaceFolder.uri;
+  const templateRootUri = vscode.Uri.joinPath(rootUri, selection.template.location);
   const edit = await createTemplateEditAsync({
     ...selection.values,
-    targetFolder: folderUri,
+    targetFolder: targetFolderUri,
     template: selection.template,
     templateRoot: templateRootUri,
-    workspace: workspaceFolder
+    workspace: targetWorkspaceFolder
   });
 
   const applied = await vscode.workspace.applyEdit(edit);
@@ -91,3 +82,39 @@ async function createNewItemsAsync(folder?: vscode.Uri | string): Promise<void> 
     throw new Error("Failed to apply template changes."); // TODO: how to get further details here?
   }
 }
+
+async function loadTemplates(workspaceFolder: vscode.WorkspaceFolder): Promise<Map<string, TemplateRootUriTuple>> {
+  const templates = new Map<string, TemplateRootUriTuple>();
+
+  // Load programatically-added templates. There is no manifest to connect them to, 
+  // so it is assumed these provide absolute or workspace-relative URIs.
+  getRegisteredTemplates().forEach((template: Template, id: string) => {
+    templates.set(id, { template: template });
+  });
+
+  // Load templates from user folder then workspace
+  const homeFolder = vscode.Uri.file(os.homedir());
+  const homeManifestUri = getManifestFileUri(homeFolder);
+  const workspaceManifestUri = getWorkspaceManifestFileUri(workspaceFolder);
+  for (var manifestUri of [homeManifestUri, workspaceManifestUri]) {
+    const manifestExists = await vscode.workspace.fs.stat(manifestUri).then(() => true, () => false);
+    if (!manifestExists) {
+      continue; // Nothing to process
+    }
+
+    const manifest = await loadManifestAsync(manifestUri);
+    manifest.templates.forEach((template: Template, index: number) => {
+      // TODO: Design ID scheme to support overriding
+      const id = `${template.name}@${manifestUri}[${index}]`;
+      templates.set(
+        id,
+        {
+          rootUri: UriUtils.dirname(manifestUri),
+          template: template,
+        });
+    });
+  }
+
+  return templates;
+}
+